+++ src/components/my/EmailBox.tsx
@@ -1,0 +1,12 @@
+import * as S from '@/styles/my/myPage.style';
+const EmailBox = () => {
+  return (
+    <>
+      <S.MainName>이메일 </S.MainName>
+      <S.DivingLine />
+      <S.SubName fontSize="15px">Clayton@gmail.com</S.SubName>
+    </>
+  );
+};
+
+export default EmailBox;
